signature_extractor/requirements.txt
ultralytics
PyMuPDF
Pillow
polars
huggingface_hub
requests
pathlib
dataclasses
typing-extensions